requirements.txt
discord.py
psycopg2
iron_cache
requests
flask
flask-admin
flask-security
flask-sqlalchemy
sqlalchemy
gunicorn